% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa.R
\docType{package}
\name{MerluzaPeruana-package}
\alias{MerluzaPeruana-package}
\alias{DYNASTIA-package}
\alias{DYNASTIA}
\title{Package for comparing 4 methods of classification (for PhD demo)}
\description{
This package was create as a demo for a PhD call and it performs a
classification by comparing 4 models (GLM, random forests, SVM and neural networks).
Version 0.0.1 includes the next functions:
\itemize{
 \item {\link{read_data}: }{Reading and preparing data for model comparing.}
 \item {\link{model_validation}: }{Take a model and returns validation results.}
}
}
\author{
Wencheng Lau-Medrano, \email{luis.laum@gmail.com}
}
\concept{comparing models}
