% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{model_validation}
\alias{model_validation}
\title{Take a model and returns validation results.}
\usage{
model_validation(model, data_test, event_level = "first", ...)
}
\arguments{
\item{model}{A model object for which validation is desired.}

\item{data_test}{A data frame in which to look for variables with which to
predict.}

\item{event_level}{A single string. Either "first" or "second" to specify
which level of truth to consider as the "event". See Details.}

\item{...}{}
}
\value{
An object of class \code{dynastia_validation}.
}
\description{
Take a model and returns validation results.
}
\details{
At version 0.0.1, \code{model} must be within next classes:
\code{glm}, \code{randomForest}, \code{svm} or \code{nnet}.
}
\examples{
purse_seine_file <- system.file("extdata", "purse_seines.rds", package = "DYNASTIA")
purse_seine <- read_data(file = purse_seine_file)

# Running model
glm_model <- glm(is_fishing ~ distance_from_port + speed + course + lat + lon + year + month,
                 family = binomial(link = "logit"), data = purse_seine$train)

model_validation(model = glm_model, data_test = purse_seine$test,
                 event_level = "second")
}
