% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{read_data}
\alias{read_data}
\title{Reading and preparing data for model comparing}
\usage{
read_data(file, split_ratio = 0.75, ...)
}
\arguments{
\item{file}{The path of file which the data are to be read from. See Detail.}

\item{split_ratio}{A number from 0 to 1 which will be used to separate data
within train and test.}

\item{...}{Extra arguments passed to main reading functions.}
}
\value{
An object of class \code{dynastia_data}.
}
\description{
Reading and preparing data for model comparing
}
\details{
\code{file} must be whether a csv or a rds file. If it is a csv, the
function will assume that it belongs to a raw data from Global Fishing Watch
datasets, particularly, from \link{https://globalfishingwatch.org/data-download/datasets/public-training-data-v1}.
}
\examples{
purse_seine_file <- system.file("extdata", "purse_seines.rds", package = "DYNASTIA")
read_data(file = purse_seine_file)
}
